+++ src/control_plane/execution_plan/mod.rs
@@ -1,0 +1,7 @@
+mod execution_plan;
+mod resolve_execution_plan;
+mod service_execution_plan;
+
+pub(crate) use execution_plan::ExecutionPlan;
+pub(crate) use resolve_execution_plan::resolve_execution_plan;
+pub(crate) use service_execution_plan::ServiceExecutionPlan;